requirements.txt
flask
openai
requests
pandas
faiss-cpu
sentence-transformers
streamlit